Python UI fixes (#1884)
* Renamed OK button; 'test sources root'

* Moved plugin settings to utbot-ui-commons
